make tests green
